backend/Python/chat_room_backend_django/requirements.txt
psycopg2-binary
django-redis
pydantic
letta